requirements.txt
fastapi==0.110.1
pydantic==2.6.4
crewai==0.50.0
crewai-tools==0.4.1
langgraph==0.0.23
uvicorn
python-dotenv
PyYAML
requests
markdown
watchdog
google-generativeai==0.7.2
openai==0.0.0
streamlit
